- Added trial binary to SConscript, fixed Win32 typing bug.
git-svn-id: https://ttuki.vtt.fi/svn/p2p-next/TUD/p2tp/trunk@541 e16421f0-f15b-0410-abcd-98678b794739

+++ exec/trial.cpp
@@ -21,7 +21,7 @@ int main (int argn, char** args) {
     p2tp::LibraryInit();
     const char* filename = "team.jpg";
     Address tracker("130.161.211.198:10000"),
-            bindaddr(INADDR_ANY,10000);
+            bindaddr((uint32_t)INADDR_ANY,10000);
     if (0>p2tp::Listen(bindaddr)) {
         print_error("cannot bind");
         return 1;
